+++ app/components/map/services/raster-map-layer-service.js
@@ -86,11 +86,11 @@ angular.module('map')
           $http.get(url).success(function (data) {
             var limits = ':' + data[0][0] + ':' + data[0][1];
             // strip existing domain if already present.
-            rasterMapLayer.wmsOptions.styles = rasterMapLayer.wmsOptions.styles.split(':')[0];
-            rasterMapLayer._imageOverlays[0].setParams({
-              styles: rasterMapLayer.wmsOptions.styles + limits
-            });
-            rasterMapLayer._imageOverlays[0].redraw();
+            // rasterMapLayer.complexWmsOptions.styles = rasterMapLayer.wmsOptions.styles.split(':')[0];
+            // rasterMapLayer._imageOverlays[0].setParams({
+            //   styles: rasterMapLayer.wmsOptions.styles + limits
+            // });
+            // rasterMapLayer._imageOverlays[0].redraw();
           });
         }
       };
@@ -103,27 +103,33 @@ angular.module('map')
       rasterMapLayer._add = function (timeState, map, options) {
 
         var opacity = options.opacity,
-            date = new Date(rasterMapLayer._mkTimeStamp(timeState.at)),
-            store = rasterMapLayer._determineStore(timeState);
-            rasterMapLayer.temporalResolution = store.resolution;
+            date = new Date(rasterMapLayer._mkTimeStamp(timeState.at));
 
-        var opts = {
-          layers: store.name,
-          styles: store.styles,
+        // Set temporalResolution for time-controller.
+        rasterMapLayer.temporalResolution = rasterMapLayer
+          .getTemporalResolution(timeState);
+
+        var defaultOptions = {
           format: 'image/png',
           version: '1.1.1',
           minZoom: 0,
           maxZoom: 21,
-          opacity: options.opacity,
-          zIndex: options.zIndex,
           crs: LeafletService.CRS.EPSG3857,
           time: _formatter(date)
         };
 
-        angular.extend(opts, rasterMapLayer.wmsOptions);
+        var opts = _.merge(defaultOptions, rasterMapLayer.complexWmsOptions);
+
+        var params = RasterService.getWmsParameters(
+          opts,
+          map.getZoom(),
+          timeState.aggWindow
+        );
+
+        params.opacity = options.opacity;
 
         rasterMapLayer._imageOverlays = [
-          LeafletService.tileLayer.wms(rasterMapLayer._imageUrlBase, opts)
+          LeafletService.tileLayer.wms(rasterMapLayer._imageUrlBase, params)
         ];
 
         rasterMapLayer._addLoadListener(
@@ -147,6 +153,7 @@ angular.module('map')
             rasterMapLayer._imageOverlays[frameIndex].setOpacity(opacity);
           }
         });
+        rasterMapLayer._opacity = opacity;
         return;
       };
 
@@ -158,7 +165,13 @@ angular.module('map')
        */
       rasterMapLayer._syncTime = function (timeState, map, wmsOptions) {
         var currentDate = rasterMapLayer._mkTimeStamp(timeState.at);
+
+        // Set temporalResolution for time-controller.
+        rasterMapLayer.temporalResolution = rasterMapLayer
+          .getTemporalResolution(timeState);
+
         rasterMapLayer._animateSyncTime(timeState, map, currentDate, wmsOptions);
+
       };
 
       /**
@@ -212,22 +225,16 @@ angular.module('map')
           // which depends on the bounds of the map and the layer and the
           // store that corresponds to the timeState.
 
-          var store = rasterMapLayer._determineStore(timeState);
-          rasterMapLayer.temporalResolution = store.resolution;
-
           var options = {
             bounds: rasterMapLayer._animationBounds,
             size: rasterMapLayer._getImageSize(map, rasterMapLayer._animationBounds)
           };
 
-          rasterMapLayer.wmsOptions.layers = store.name;
-          rasterMapLayer.wmsOptions.styles = store.styles;
-
           rasterMapLayer.url = RasterService.buildURLforWMS(
             rasterMapLayer._imageUrlBase,
             map,
             timeState.playing,
-            rasterMapLayer.wmsOptions,
+            rasterMapLayer.complexWmsOptions,
             options
           );
 
@@ -330,50 +337,14 @@ angular.module('map')
        * Local helper that returns a rounded timestamp
        */
       rasterMapLayer._mkTimeStamp = function (t) {
-        var result = UtilService.roundTimestamp(t, rasterMapLayer.temporalResolution, false);
+        var result = UtilService.roundTimestamp(t, rasterMapLayer.getTemporalResolution(), false);
         return result;
       };
 
-      /**
-       * @description based on the temporal window. The time between
-       * timestate.start and timestate.end determines which store is to be used.
-       * This only works for radar stuff.
-       */
-      rasterMapLayer._determineStore = function (timeState) {
-
-        if (rasterMapLayer.slug !== 'radar/5min') {
-          return {
-            name: rasterMapLayer.slug,
-            styles: rasterMapLayer.wmsOptions.styles,
-            resolution: rasterMapLayer.temporalResolution
-          };
-        }
-
-        var resolutionHours = (timeState.aggWindow) / 60 / 60 / 1000;
-
-        var aggType = [rasterMapLayer.slug.split('/')[0]];
-
-        if (resolutionHours >= 24) {
-          aggType[1] = 'day';
-        } else if (resolutionHours >= 1 && resolutionHours < 24) {
-          aggType[1] = 'hour';
-        } else {
-          aggType[1] = '5min';
-        }
-        var resolutions = {
-          '5min': 300000,
-          'hour': 3600000,
-          'day': 86400000
-        };
-
-        return {
-          styles: aggType.join('-'),
-          name: aggType.join('/'),
-          resolution: resolutions[aggType[1]]
-        };
-
+      rasterMapLayer.getTemporalResolution = function (timeState) {
+        var resolution = rasterMapLayer.minFrequency;
+        return resolution;
       };
-
 
       /**
        * @description Removes old frame by looking for a frame that has an
@@ -396,7 +367,7 @@ angular.module('map')
 
         var newFrame = rasterMapLayer._imageOverlays[currentOverlayIndex];
         // Turn on new frame
-        newFrame.setOpacity(wmsOptions.opacity);
+        newFrame.setOpacity(rasterMapLayer._opacity);
       };
 
       /**