+++ app/components/map/services/wms-map-layer-service.js
@@ -8,23 +8,43 @@
  * Additional methods used to extend nxtLayer with leaflet/map specific methods.
  */
 angular.module('map')
-.factory('wmsMapLayer', ['$q', 'LeafletService', 'MapLayerService',
-  function ($q, LeafletService, MapLayerService) {
+.factory('wmsMapLayer', [
+  '$q',
+  'LeafletService',
+  'MapLayerService',
+  'RasterService',
+  function ($q, LeafletService, MapLayerService, RasterService) {
 
     return function (options) {
 
       var wmsMapLayer = options;
 
-      wmsMapLayer.wms = MapLayerService.createWmsLayer(options);
+      var params = RasterService.getWmsParameters(
+        wmsMapLayer.complexWmsOptions,
+        0,
+        0
+      );
+
+      wmsMapLayer.wms = MapLayerService.createWmsLayer(
+        _.extend(options, params)
+      );
 
       wmsMapLayer.update = function (map, timeState, options) {
         var promise;
 
         wmsMapLayer.wms.setOpacity(options.opacity);
 
+        var params = RasterService.getWmsParameters(
+          wmsMapLayer.complexWmsOptions,
+          map.getZoom(),
+          timeState.aggWindow
+        );
+
         if (!map.hasLayer(wmsMapLayer.wms)) {
           map.addLayer(wmsMapLayer.wms);
         }
+
+        wmsMapLayer.wms.setParams(params);
 
         return promise;
       };